midi_translator/requirements.txt
usb-device-midi
